requirements.txt
torch
torchvision
torchaudio

vllm

transformers
tokenizers


sentence-transformers

lightgbm
xgboost

scikit-learn
numpy
pandas


scipy

neo4j

chromadb

langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=1.0.0

openai

# 通义千问API（用于Agent对话和JD生成，免费200万token）
dashscope


fastapi
uvicorn[standard]
pydantic
python-multipart


ujson>=5.8.0

pyyaml


python-dateutil


requests
httpx

aiohttp


beautifulsoup4
lxml


drissionpage


tqdm


loguru


python-dotenv

pydantic-settings


pytest

jieba

py2neo

# MySQL数据库相关
SQLAlchemy>=2.0.0
PyMySQL>=1.0.0
alembic>=1.0.0

# JWT认证相关
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
